+++ src/components/Home/functions/locationToID.js
@@ -1,0 +1,8 @@
+function locationToID(location) {
+    let x = location.x;
+    let y = location.y;
+    x = String.fromCharCode(65 + x - 1);
+    return x + y;
+}
+
+export default locationToID;