+++ src/main.jsx
@@ -4,11 +4,7 @@ import App from './App.jsx'
 import './index.css'
 
 import {BrowserRouter} from 'react-router-dom'
-import Game from './models/game.js'
 
-
-const game = new Game(100);
-game.startGame();
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>