+++ src/models/pieces/bishop.js
@@ -10,14 +10,14 @@ class Bishop extends Piece {
         super(color, position, asset, Constants.PIECE_TYPE_BISHOP);
     }
 
-    getPossibleMoves(board, opponentPieces, playerPieces) {
+    getPossibleMoves(opponentPieces, playerPieces) {
         let possibleMoves = [];
         let x = this.position.x;
         let y = this.position.y;
 
         // Check if the bishop can move up-right
-        for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x + i, y + i, opponentPieces, playerPieces);
+        for (let i = 1; i < 8; i++) {
+            let status = this.cellStatus(x + i, y + i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x + i, y: y + i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -29,8 +29,8 @@ class Bishop extends Piece {
         }
 
         // Check if the bishop can move up-left
-        for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x - i, y + i, opponentPieces, playerPieces);
+        for (let i = 1; i < 8; i++) {
+            let status = this.cellStatus(x - i, y + i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x - i, y: y + i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -42,8 +42,8 @@ class Bishop extends Piece {
         }
 
         // Check if the bishop can move down-right
-        for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x + i, y - i, opponentPieces, playerPieces);
+        for (let i = 1; i < 8; i++) {
+            let status = this.cellStatus(x + i, y - i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x + i, y: y - i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -55,8 +55,8 @@ class Bishop extends Piece {
         }
 
         // Check if the bishop can move down-left
-        for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x - i, y - i, opponentPieces, playerPieces);
+        for (let i = 1; i < 8; i++) {
+            let status = this.cellStatus(x - i, y - i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x - i, y: y - i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {