+++ src/models/pieces/piece.js
@@ -13,8 +13,8 @@ class Piece {
         console.log(`${this.color} ${this.type} moved to ${String.fromCharCode(position.x + 64)}${position.y}`);
     }
 
-    cellStatus(board, x, y, opponentPieces, playerPieces = []) {
-        if (board[y][x] === Constants.BOARD_COLOR_BORDER || x < 1 || x > 8 || y < 1 || y > 8) {
+    cellStatus(x, y, opponentPieces, playerPieces) {
+        if (x < 1 || x > 8 || y < 1 || y > 8) {
             return Constants.CELL_STATUS_OUT_OF_BOUNDS;
         }
         if (opponentPieces.some(piece => piece.position.x === x && piece.position.y === y)) {
@@ -27,8 +27,8 @@ class Piece {
         return Constants.CELL_STATUS_EMPTY;
     }
 
-    getPossibleMoves(board, opponentPieces) {
-        throw new Error("Method 'getPossibleMoves(board, opponentPieces)' must be implemented.");
+    getPossibleMoves(opponentPieces, playerPieces) {
+        throw new Error("Method 'getPossibleMoves(opponentPieces, playerPieces)' must be implemented.");
     }
 
     getLocationCode() {