+++ src/models/pieces/queen.js
@@ -10,14 +10,14 @@ class Queen extends Piece {
         super(color, position, asset, Constants.PIECE_TYPE_QUEEN);
     }
 
-    getPossibleMoves(board, opponentPieces, playerPieces) {
+    getPossibleMoves(opponentPieces, playerPieces) {
         let possibleMoves = [];
         let x = this.position.x;
         let y = this.position.y;
 
         // Check if the queen can move up
         for (let i = y + 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x, i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x, i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x, y: i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -30,7 +30,7 @@ class Queen extends Piece {
 
         // Check if the queen can move down
         for (let i = y - 1; i >= 1; i--) {
-            let status = this.cellStatus(board, x, i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x, i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x, y: i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -43,7 +43,7 @@ class Queen extends Piece {
 
         // Check if the queen can move left
         for (let i = x - 1; i >= 1; i--) {
-            let status = this.cellStatus(board, i, y, opponentPieces, playerPieces);
+            let status = this.cellStatus(i, y, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: i, y: y});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -56,7 +56,7 @@ class Queen extends Piece {
 
         // Check if the queen can move right
         for (let i = x + 1; i <= 8; i++) {
-            let status = this.cellStatus(board, i, y, opponentPieces, playerPieces);
+            let status = this.cellStatus(i, y, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: i, y: y});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -69,7 +69,7 @@ class Queen extends Piece {
 
         // Check if the queen can move up-right
         for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x + i, y + i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x + i, y + i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x + i, y: y + i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -82,7 +82,7 @@ class Queen extends Piece {
 
         // Check if the queen can move up-left
         for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x - i, y + i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x - i, y + i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x - i, y: y + i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -95,7 +95,7 @@ class Queen extends Piece {
 
         // Check if the queen can move down-right
         for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x + i, y - i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x + i, y - i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x + i, y: y - i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -108,7 +108,7 @@ class Queen extends Piece {
 
         // Check if the queen can move down-left
         for (let i = 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x - i, y - i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x - i, y - i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x - i, y: y - i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {