+++ src/models/pieces/rook.js
@@ -10,14 +10,14 @@ class Rook extends Piece {
         super(color, position, asset, Constants.PIECE_TYPE_ROOK);
     }
 
-    getPossibleMoves(board, opponentPieces, playerPieces) {
+    getPossibleMoves(opponentPieces, playerPieces) {
         let possibleMoves = [];
         let x = this.position.x;
         let y = this.position.y;
 
         // Check if the rook can move up
         for (let i = y + 1; i <= 8; i++) {
-            let status = this.cellStatus(board, x, i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x, i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x, y: i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -30,7 +30,7 @@ class Rook extends Piece {
 
         // Check if the rook can move down
         for (let i = y - 1; i >= 1; i--) {
-            let status = this.cellStatus(board, x, i, opponentPieces, playerPieces);
+            let status = this.cellStatus(x, i, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: x, y: i});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -43,7 +43,7 @@ class Rook extends Piece {
 
         // Check if the rook can move left
         for (let i = x - 1; i >= 1; i--) {
-            let status = this.cellStatus(board, i, y, opponentPieces, playerPieces);
+            let status = this.cellStatus(i, y, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: i, y: y});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {
@@ -56,7 +56,7 @@ class Rook extends Piece {
 
         // Check if the rook can move right
         for (let i = x + 1; i <= 8; i++) {
-            let status = this.cellStatus(board, i, y, opponentPieces, playerPieces);
+            let status = this.cellStatus(i, y, opponentPieces, playerPieces);
             if (status === Constants.CELL_STATUS_EMPTY) {
                 possibleMoves.push({x: i, y: y});
             } else if (status === Constants.CELL_STATUS_OCCUPIED_OPPONENT) {