+++ src/pages/Home.jsx
@@ -2,8 +2,11 @@ import "../common.css"
 import "./styles/Home.css"
 import HomeBoard from "../components/Home/Board"
 import HomeHistory from "../components/Home/History"
+import { game } from "../models/variables"
 
 function HomePage() {
+  game.startGame();
+
   return (
     <>
     <div className="screen">